# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Satori R1.1 API
  version: 1.1.0
  description: R1.0 API plus the R1.1 commerce, entitlement, consumption and membership candidate contract frozen by ADR-0003.
servers:
  - url: https://test-satori.shenxinyou.com/api/v1
    description: R1.0 test environment
  - url: http://localhost:3000/api/v1
    description: Local API
...
paths:
...
  /me:
    get:
      tags: [Me]
      operationId: getMe
      responses:
        '200': {description: Current user routing state, content: {application/json: {schema: {$ref: '#/components/schemas/MeEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /me/registration-reward:
    get:
      tags: [WisdomSeed]
      operationId: getRegistrationReward
      responses:
        '200': {description: Registration reward eligibility and status, content: {application/json: {schema: {$ref: '#/components/schemas/RegistrationRewardEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/registration-reward/claim:
    post:
      tags: [WisdomSeed]
      operationId: claimRegistrationReward
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses:
        '200': {description: Reward claimed or original claim replayed, content: {application/json: {schema: {$ref: '#/components/schemas/RegistrationRewardClaimEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/wisdom-seed-account:
    get:
      tags: [WisdomSeed]
      operationId: getWisdomSeedAccount
      responses:
        '200': {description: Wisdom seed account, content: {application/json: {schema: {$ref: '#/components/schemas/WisdomSeedAccountEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/wisdom-seed-transactions:
    get:
      tags: [WisdomSeed]
      operationId: listWisdomSeedTransactions
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Wisdom seed ledger, content: {application/json: {schema: {$ref: '#/components/schemas/WisdomSeedTransactionListEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    refreshCookie: {type: apiKey, in: cookie, name: satori_refresh}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    Cursor: {name: cursor, in: query, required: false, schema: {type: string}}
    Limit: {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
...
  responses:
    Error:
      description: API v1 error
      headers: {X-Request-Id: {schema: {type: string}}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, retryable: {type: boolean}, nextAction: {type: [string, 'null']}, retryAfterSeconds: {type: [integer, 'null'], minimum: 0}, details: {type: [object, 'null'], additionalProperties: true}}
    PageMeta: {type: object, required: [nextCursor, hasMore], properties: {nextCursor: {type: [string, 'null']}, hasMore: {type: boolean}}}
    Preferences: {type: object, required: [timezone, locale], properties: {timezone: {type: string, examples: [Asia/Shanghai]}, timezoneSource: {type: string}, locale: {type: string, examples: [zh-CN]}, updatedAt: {type: string, format: date-time}}}
...
    LegalDocumentSummary: {type: object, required: [documentId, type, version, title, required], properties: {documentId: {type: string}, type: {type: string, enum: [PRIVACY_POLICY, TERMS_OF_SERVICE, AI_CONTENT_NOTICE]}, version: {type: string}, title: {type: string}, required: {type: boolean}, publishedAt: {type: string, format: date-time}}}
...
    UserSummary: {type: object, required: [userId, status, phoneMasked, requiresConsent, createdAt], properties: {userId: {type: string}, status: {type: string, enum: [ACTIVE, DELETION_PENDING, DELETING, DISABLED]}, phoneMasked: {type: string}, requiresConsent: {type: boolean}, createdAt: {type: string, format: date-time}}}
    NextAction: {type: string, enum: [ACCEPT_CONSENTS, CREATE_PROFILE, CONFIRM_PROFILE, CLAIM_REGISTRATION_REWARD, CREATE_TODAY_DAILY_INSIGHT, VIEW_HOME]}
...
    MeEnvelope: {type: object, required: [data], properties: {data: {allOf: [{$ref: '#/components/schemas/UserSummary'}, {type: object, required: [preferences, nextAction], properties: {preferences: {$ref: '#/components/schemas/Preferences'}, pendingConsents: {type: array, items: {$ref: '#/components/schemas/LegalDocumentSummary'}}, profileState: {type: string, enum: [NOT_CREATED, CALCULATED, ACTIVE]}, nextAction: {$ref: '#/components/schemas/NextAction'}}}]}}}
...
    WisdomSeedAccount: {type: object, required: [accountId, available, reserved, totalEarned, totalSpent, updatedAt], properties: {accountId: {type: string}, available: {type: integer, minimum: 0}, reserved: {type: integer, minimum: 0}, totalEarned: {type: integer, minimum: 0}, totalSpent: {type: integer, minimum: 0}, updatedAt: {type: string, format: date-time}}}
    WisdomSeedAccountEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/WisdomSeedAccount'}}}
    WisdomSeedTransaction: {type: object, required: [transactionId, type, amount, balanceAfter, businessType, resourceId, createdAt], properties: {transactionId: {type: string}, type: {type: string, enum: [GRANT, RESERVE, CONSUME, RELEASE, REFUND, ADJUSTMENT]}, amount: {type: integer}, balanceAfter: {type: integer, minimum: 0}, businessType: {type: string, enum: [REGISTRATION_REWARD, DAILY_INSIGHT]}, resourceId: {type: string}, originalTransactionId: {type: [string, 'null']}, title: {type: string}, createdAt: {type: string, format: date-time}}}
    WisdomSeedTransactionListEnvelope: {type: object, required: [data, meta], properties: {data: {type: array, items: {$ref: '#/components/schemas/WisdomSeedTransaction'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
    RegistrationReward: {type: object, required: [rewardId, rewardType, status, wisdomSeedAmount], properties: {rewardId: {type: string}, rewardType: {const: NEW_USER_ONBOARDING}, status: {type: string, enum: [AVAILABLE, CLAIMED, EXPIRED, INELIGIBLE]}, wisdomSeedAmount: {type: integer, minimum: 0}, claimedAt: {type: [string, 'null'], format: date-time}, expiresAt: {type: [string, 'null'], format: date-time}}}
    RegistrationRewardEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/RegistrationReward'}}}
    RegistrationRewardClaimEnvelope: {type: object, required: [data], properties: {data: {type: object, required: [reward, account, transaction], properties: {reward: {$ref: '#/components/schemas/RegistrationReward'}, account: {$ref: '#/components/schemas/WisdomSeedAccount'}, transaction: {$ref: '#/components/schemas/WisdomSeedTransaction'}}}}}
...
security:
  - bearerAuth: []